Nasty routing error fixed
The CH5 and CH6 did not work. That was because of one
unobserved airwire. The ground was not routed to the IC4. Too bad
On the boards the airwire must be mounted connecting the ground on C5
with the pin 4 of the IC4

The .brd is updated with one extra via, but because the boards are
already manufactured, there is nothing that can be redone. Solder a
wire...