Saving forced states
this is primarily to sort out storing the forced states of the
ADCs (FORCE ON, FORCE OFF and NORMAL). I've decided to only store them in
EEPROM. Whenever the EEPROM config is valid, the force state is used
even if the SD config has changed.

The config.txt has got a header with neccesary explanations. The parser somewhat
improved (but still simple)